VIC-1617 Victor BLE: Use local copy of strlcpy

+++ external/anki-ble/ankibluetoothd/btstack.cpp
@@ -16,8 +16,8 @@
 #include "log.h"
 #include "btutils.h"
 #include "gatt_constants.h"
+#include "strlcpy.h"
 #include "taskExecutor.h"
-#include <cutils/memory.h>
 #include <hardware/bluetooth.h>
 #include <hardware/bt_gatt.h>
 #include <hardware/bt_gatt_client.h>
@@ -1412,7 +1412,7 @@ bool SetAdapterName(const std::string& name) {
          name.c_str(), name.size(), sizeof(bd_name.name) - 1);
     return false;
   }
-  bt_property_t property = {BT_PROPERTY_BDNAME, strlen((char *) bd_name.name), &bd_name};
+  bt_property_t property = {BT_PROPERTY_BDNAME, (int) strlen((char *) bd_name.name), &bd_name};
   int status = sBtInterface->set_adapter_property(&property);
   if (status != BT_STATUS_SUCCESS) {
     loge("failed to set adapter name to %s. status = %s",
